combined usd layer with isoTP layer ATTENTION not working properly. UDS layer closes isoTP too soon but very first message is received and answered correctly

+++ MCU_Aurix/bootloader/inc/isotp.h
@@ -56,7 +56,7 @@ isoTP* isotp_init(void);
 
 void isotp_send(isoTP* iso, uint8_t* data, uint32_t data_in_len);
 
-uint8_t* isotp_rcv(int16_t* total_length);
+uint8_t* isotp_rcv(uint32_t* total_length);
 
 void process_can_to_isotp(uint32_t* rxData, IfxCan_DataLengthCode dlc);
